pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sitl"
version = "0.0.5"
description = "A SITL for Betaflight"
readme = "README.md"
authors = [{name = "Jonas Eschmann", email = "jonas.eschmann@gmail.com"}]
license = {text = "MIT"}
dependencies = [
  "betaflight",
  "l2f",
  "numpy",
  "scipy",
  "pygame",
  "websockify",
  "rerun-sdk"
]


[tool.hatch.build.targets.wheel]
packages = ["sitl"]

[project.scripts]
sitl = "sitl.main:sync_main"
sitl-websockify = "sitl.websockify_proxy:main"
sitl-gamepad = "sitl.gamepad:main"

